requirements.txt
networkx
numpy
pandas
pytest
pyyaml
requests
scipy
sqlalchemy
statsmodels
tabula
tiktoken
